pyproject.toml
[tool.poetry]
name = "twomoons"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.10.2"
numpy = "^1.22.2"
scikit-learn = "^1.0.2"
laplace-torch = "^0.1a2"
matplotlib = "^3.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
